Separate and refactor controllers
- Split controllers into separate modules.
- Implemented centralized error handling:
  - Removed unnecessary try/catch blocks in controllers.
  - Added the 'error' hook to koa.
  - Now, every error that was thrown in a controller is handled by
    api/error.js:`processError`.
- Worked on isolation of the API server from the rest of the app.
