+++ src/api/utils/users.js
@@ -1,0 +1,27 @@
+/*
+ This file is a part of libertysoil.org website
+ Copyright (C) 2015  Loki Education (Social Enterprise)
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+export async function areMutuallyFollowed(knex, user1Id, user2Id) {
+  const userFollows = await knex('followers')
+    .where({ user_id: user1Id, following_user_id: user2Id })
+    .count();
+  const userBeingFollowed = await knex('followers')
+    .where({ user_id: user2Id, following_user_id: user1Id })
+    .count();
+
+  return parseInt(userFollows[0].count, 10) > 0 && parseInt(userBeingFollowed[0].count, 10) > 0;
+}
